+++ server/src/index.js
@@ -1,11 +1,12 @@
+import * as Middleware from './middleware'
 
 import Express from 'express'
-import configureRoutes from './configureRoutes'
 
 const app = Express()
 
 app.set('view engine', 'pug')
-app.use(configureRoutes)
+app.use(Middleware.routes)
+app.use(Middleware.api)
 app.listen(2001, () => {
   console.log('server is running.')
 })
