+++ server/src/middleware/api.js
@@ -1,0 +1,7 @@
+
+export const Api = async (req, res, next) => {
+  console.log('This is API handlers')
+  next()
+}
+
+export default Api
