+++ server/src/middleware/index.js
@@ -1,0 +1,10 @@
+
+import routes from './routes'
+import api from './api'
+import socket from './socket'
+
+export {
+  routes,
+  api,
+  socket
+}
