+++ server/src/middleware/socket.js
@@ -1,0 +1,4 @@
+
+export default (req, res, next) => {
+
+}
